change from long to int

+++ src/ex2_1/CounterThread.java
@@ -3,23 +3,24 @@ package ex2_1;
 import java.io.BufferedReader;
 import java.io.FileReader;
 import java.io.IOException;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 public class CounterThread extends Thread {
 
     private final String fileName;
 
-    private AtomicLong counter;
+    private AtomicInteger counter;
 
     public CounterThread (String fileName){
 
         if (fileName == null) throw new NullPointerException("insert valid file name");
 
         this.fileName = fileName;
-        counter = new AtomicLong();
+        counter = new AtomicInteger();
     }
 
-    public Long getCounter (){
+    public int getCounter (){
         return counter.get();
     }
